Consolidate fast and slow path tests into test_skvbc_commit_path.
